requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
apify-client==1.3.1
apify-shared==1.0.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.0
constantly==15.1.0
contourpy==1.1.0
cryptography==41.0.3
cssselect==1.2.0
cycler==0.11.0
datasets==2.14.1
dill==0.3.7
exceptiongroup==1.1.3
filelock==3.12.2
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.57.0
h11==0.14.0
h5py==3.9.0
haversine==2.8.0
httpcore==0.17.3
httpx==0.24.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.6
lit==16.0.6
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
openai==0.27.8
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parsel==1.8.1
Pillow==10.0.0
Protego==0.3.0
protobuf==4.24.1
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyDispatcher==2.0.7
pyOpenSSL==23.2.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
queuelib==1.6.2
regex==2023.6.3
replicate==0.11.0
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.2
Scrapy==2.10.0
selenium==4.11.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
threadpoolctl==3.2.0
tldextract==3.4.4
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.31.0
trio==0.22.2
trio-websocket==0.10.3
triton==2.0.0
Twisted==22.10.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
w3lib==2.1.2
webdriver-manager==4.0.0
Werkzeug==2.3.7
wrapt==1.15.0
wsproto==1.2.0
xxhash==3.3.0
yarl==1.9.2
zope.interface==6.0
google-cloud-storage==2.10.0